Refactor : remove handleDate.tsx
